Organised and cleaned authentication files
