+++ src/app/shared/authentication/IUser.ts
