+++ src/app/shared/authentication/authentication.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 
 import { AuthenticationService } from './authentication.service';
 import { AngularFireModule } from '@angular/fire';
-import { environment } from '../../environments/environment';
+import { environment } from '../../../environments/environment';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('AuthenticationService', () => {
