+++ src/app/sign-in/sign-in.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SignInComponent } from './sign-in.component';
-import { AuthenticationService } from '../authentication/authentication.service';
+import { AuthenticationService } from '../shared/authentication/authentication.service';
 import { AngularFireModule } from '@angular/fire';
 import { environment } from '../../environments/environment';
 import { RouterTestingModule } from '@angular/router/testing';
