+++ src/app/sign-in/sign-in.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AuthenticationService } from '../authentication/authentication.service';
+import { AuthenticationService } from '../shared/authentication/authentication.service';
 import { FormControl } from '@angular/forms';
 
 @Component({
